Add bulk renaming function for VM interfaces
