+++ netbox/templates/utilities/obj_bulk_rename.html
