+++ netbox/virtualization/urls.py
@@ -55,6 +55,7 @@ urlpatterns = [
     path('interfaces/add/', views.InterfaceCreateView.as_view(), name='vminterface_add'),
     path('interfaces/import/', views.InterfaceBulkImportView.as_view(), name='vminterface_import'),
     path('interfaces/edit/', views.InterfaceBulkEditView.as_view(), name='vminterface_bulk_edit'),
+    path('interfaces/rename/', views.InterfaceBulkRenameView.as_view(), name='vminterface_bulk_rename'),
     path('interfaces/delete/', views.InterfaceBulkDeleteView.as_view(), name='vminterface_bulk_delete'),
     path('interfaces/<int:pk>/', views.InterfaceView.as_view(), name='vminterface'),
     path('interfaces/<int:pk>/edit/', views.InterfaceEditView.as_view(), name='vminterface_edit'),
